+++ test/gas/ForwarderRequestCancel.t.sol
@@ -192,7 +192,7 @@ contract ForwarderRequestCancelTest is Test {
             to: to,
             paymentToken: _paymentToken,
             data: data,
-            deadline: block.timestamp + 30 days,
+            deadline: uint64(block.timestamp + 30 days),
             nonce: nonce,
             signature: abi.encodePacked(r2, s2, v2)
         });
